+++ source/WindowsAPICodePack/Win32Native.Shared/APIs/Shell/Enums.cs
@@ -2,6 +2,39 @@
 
 namespace Microsoft.WindowsAPICodePack.Win32Native.Shell
 {
+    public enum ShellImageListIconSize
+    {
+        /// <summary>
+        /// The image size is normally 32x32 pixels. However, if the Use large icons option is selected from the Effects section of the Appearance tab in Display Properties, the image is 48x48 pixels.
+        /// </summary>
+        Large = 0,
+
+        /// <summary>
+        /// These images are the Shell standard small icon size of 16x16, but the size can be customized by the user.
+        /// </summary>
+        Small = 1,
+
+        /// <summary>
+        /// These images are the Shell standard extra-large icon size. This is typically 48x48, but the size can be customized by the user.
+        /// </summary>
+        ExtraLarge = 2,
+
+        /// <summary>
+        /// These images are the size specified by GetSystemMetrics called with SM_CXSMICON and GetSystemMetrics called with SM_CYSMICON.
+        /// </summary>
+        SysSmall = 3,
+
+        /// <summary>
+        /// Windows Vista and later. The image is normally 256x256 pixels.
+        /// </summary>
+        Jumbo = 4,
+
+        /// <summary>
+        /// The largest valid flag value, for validation purposes.
+        /// </summary>
+        Last = Jumbo
+    }
+
     public enum NotifyIconVersion : uint
     {
         None = 0,